Move oracle and evaluation scale protocols into chipiron.core
