+++ src/chipiron/core/evaluation_scale.py
