+++ src/chipiron/players/factory_pipeline.py
@@ -18,7 +18,7 @@ from chipiron.players.move_selector.move_selector_args import (
     NonTreeMoveSelectorArgs,
 )
 from chipiron.players.move_selector.tree_and_value_args import TreeAndValueAppArgs
-from chipiron.players.oracles import PolicyOracle, TerminalOracle, ValueOracle
+from chipiron.core.oracles import PolicyOracle, TerminalOracle, ValueOracle
 from chipiron.players.player import GameAdapter, Player
 from chipiron.players.player_args import HasMoveSelectorType
 from chipiron.scripts.chipiron_args import ImplementationArgs
