+++ tests/checkers/test_checkers_terminal_scale.py
@@ -8,7 +8,7 @@ pytest.importorskip("valanga")
 from valanga import Color
 
 from chipiron.games.domain.game.game_rules import GameOutcome, OutcomeKind
-from chipiron.players.boardevaluators.evaluation_scale import (
+from chipiron.core.evaluation_scale import (
     EvaluationScale,
     get_value_over_enum,
 )
